requirements.txt
fastapi
uvicorn
pydantic
pathlib
python-dotenv
fastapi-mcp
pip-system-certs